backend/requirements.txt
# Core data science packages
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0

# UMAP for dimensionality reduction
umap-learn>=0.5.0,<1.0.0

# HDBSCAN for clustering
hdbscan>=0.8.0,<1.0.0

# Gensim for text processing and word embeddings
gensim>=4.0.0,<5.0.0

# MongoDB driver
pymongo>=4.0.0,<5.0.0

# Environment variable management
python-dotenv>=0.19.0,<1.0.0

# Additional dependencies that might be needed
scipy>=1.7.0,<2.0.0
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0

# FastAPI and related packages (if using the backend API)
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-multipart>=0.0.5,<1.0.0
